Lobby now knows up-to-date gameData and calculates user's playerId and updates gameData on player disconnection
